// entidad-web/entidad-web/src/app/@data/model/cabeceraMaestraDetail.ts
export class CabeceraMaestraDetail {
    estadoRegistro: string;
    usuarioCreacion: string;
    fechaCreacion: string;
    usuarioModificacion: string;
    fechaModificacion: string;
    maeDetalleId: number;
    maeCabeceraId: number;
    descripcion: string;
    descripcionCorta: string;
    sigla: string;
    referencia: string;
    orden: string;
    textToSearch?: string;
    descripcionMaestra?: string;
    settings?: any;
}